# api/openapi.yaml
openapi: 3.1.0
info:
  title: Memora Backend API
  description: Offline-first, consent-gated cognitive-stimulation and memory-assistance API for SIH26003 (MDoNER).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /v1/health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: 1.0.0

  /v1/patients:
    get:
      summary: List patients accessible by the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
    post:
      summary: Register a new patient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'

  /v1/patients/{patient_id}:
    get:
      summary: Get patient details
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          description: Not authorized to access this patient
        '404':
          description: Patient not found

  /v1/patients/{patient_id}/ability:
    get:
      summary: Get longitudinal ability score history for a patient
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of recorded ability scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbilityScore'
        '403':
          description: Not authorized to access this patient

  /v1/onboarding:
    post:
      summary: Onboard a new elder with capacity evaluation and verifiable guardian consent
      description: Atomically creates patient profile and records guardian consent under DPDP Act 2023. Patient data remains locked until valid consent is recorded.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '201':
          description: Onboarding complete and consent established
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingResponse'
        '400':
          description: Invalid onboarding payload

  /v1/consents:
    post:
      summary: Record verifiable consent for a patient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentCreate'
      responses:
        '201':
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '403':
          description: Not authorized to grant consent for this patient

  /v1/events:batch:
    post:
      summary: Ingest a batch of offline game events (idempotent, consent-gated)
      description: Ingests up to 200 client-generated UUIDv7 game events. Requires active consent. Duplicate event IDs are ignored without error.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchEventsRequest'
      responses:
        '200':
          description: Batch processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchEventsResponse'
        '400':
          description: Invalid batch or batch exceeds 200 events
        '403':
          description: Consent required before writing patient activity data

  /v1/patients/{patient_id}/memories:
    get:
      summary: List private memories for an elder (authorized members only)
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [photo, song, story]
      responses:
        '200':
          description: List of memories with time-limited signed URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
        '403':
          description: Not authorized or consent missing
        '404':
          description: Patient not found
    post:
      summary: Upload a new private memory (photo, song, or story)
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreate'
      responses:
        '201':
          description: Memory saved with signed URL and vector embedding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '403':
          description: Not authorized or consent missing

  /v1/memories/{memory_id}:
    get:
      summary: Retrieve a single memory with a fresh signed URL
      security:
        - BearerAuth: []
      parameters:
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory details with signed URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '403':
          description: Not authorized
        '404':
          description: Memory not found
    delete:
      summary: Permanently delete a memory, its private storage file, and vector embedding
      security:
        - BearerAuth: []
      parameters:
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory, storage asset, and embedding permanently deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  id:
                    type: string
                  file_deleted:
                    type: boolean
                  embedding_deleted:
                    type: boolean
        '403':
          description: Not authorized
        '404':
          description: Memory not found

  /v1/patients/{patient_id}/approval-queue:
    get:
      summary: Retrieve pending AI-generated quiz questions awaiting caregiver review
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of pending quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeneratedQuizItem'
        '403':
          description: Not authorized or consent missing

  /v1/quiz-items/{item_id}/approve:
    post:
      summary: Approve an AI-generated quiz question for patient activity sessions
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question approved and unlocked for patient play
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedQuizItem'
        '403':
          description: Not authorized
        '404':
          description: Quiz item not found

  /v1/quiz-items/{item_id}/reject:
    post:
      summary: Reject an AI-generated quiz question (guaranteed never to be served to patient)
      security:
        - BearerAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedQuizItem'
        '403':
          description: Not authorized
        '404':
          description: Quiz item not found

  /v1/patients/{patient_id}/quiz-items:
    get:
      summary: Retrieve approved quiz items for patient sessions and offline sync (strictly excludes unapproved)
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of approved quiz items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeneratedQuizItem'
        '403':
          description: Not authorized or consent missing

  /v1/asha/patients:
    get:
      summary: Retrieve patients assigned to the authenticated ASHA worker, ordered by triage priority
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of assigned patients ordered by triage need
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AshaPatientSummary'
        '403':
          description: Not authorized as ASHA or clinician

  /v1/asha/patients/{patient_id}:
    get:
      summary: Retrieve patient details and anomaly evidence for assigned ASHA
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient triage details and longitudinal evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AshaPatientDetail'
        '403':
          description: Patient not assigned to this ASHA worker
        '404':
          description: Patient not found

  /v1/asha/patients/{patient_id}/report.pdf:
    get:
      summary: Generate server-side non-diagnostic clinical summary PDF report with evidence
      security:
        - BearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Generated PDF report binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Patient not assigned to this ASHA worker
        '404':
          description: Patient not found


components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Patient:
      type: object
      required:
        - id
        - full_name
        - preferred_language
        - current_theta
        - created_at
      properties:
        id:
          type: string
          example: 018f3a2b-8c4d-7000-8000-000000000001
        full_name:
          type: string
          example: Synthetic Patient A
        preferred_language:
          type: string
          example: as
        current_theta:
          type: number
          format: float
          example: 0.25
        created_at:
          type: string
          format: date-time

    PatientCreate:
      type: object
      required:
        - full_name
        - preferred_language
      properties:
        full_name:
          type: string
          example: Synthetic Patient A
        preferred_language:
          type: string
          enum: [en, hi, as]
          example: as
        initial_theta:
          type: number
          format: float
          default: 0.0

    Consent:
      type: object
      required:
        - id
        - patient_id
        - guardian_name
        - guardian_relationship
        - consent_version
        - is_active
        - granted_at
      properties:
        id:
          type: string
        patient_id:
          type: string
        guardian_name:
          type: string
        guardian_relationship:
          type: string
        guardian_note:
          type: string
        capacity_assessed:
          type: boolean
        capacity_notes:
          type: string
        consent_version:
          type: string
        is_active:
          type: boolean
        granted_at:
          type: string
          format: date-time

    ConsentCreate:
      type: object
      required:
        - patient_id
        - guardian_name
        - guardian_relationship
        - consent_version
      properties:
        patient_id:
          type: string
        guardian_name:
          type: string
          example: Ananya Goswami
        guardian_relationship:
          type: string
          example: Daughter
        guardian_note:
          type: string
          example: Registered primary caregiver; elder requires full digital assistance
        capacity_assessed:
          type: boolean
          example: true
        capacity_notes:
          type: string
          example: Cognitive decline reported; family consensus
        consent_version:
          type: string
          example: "2026.1"

    OnboardingRequest:
      type: object
      required:
        - full_name
        - preferred_language
        - lacks_capacity
        - guardian_name
        - guardian_relationship
        - consent_version
      properties:
        full_name:
          type: string
          example: Bhaben Baruah
        preferred_language:
          type: string
          example: as
        initial_theta:
          type: number
          format: float
          default: 0.0
        lacks_capacity:
          type: boolean
          description: Does the elder lack independent legal capacity to consent?
          example: true
        capacity_notes:
          type: string
          example: Memory decline observed over 12 months; daughter assists with all care decisions
        guardian_name:
          type: string
          example: Jonali Baruah
        guardian_relationship:
          type: string
          example: Daughter
        guardian_note:
          type: string
          example: Primary caregiver residing with elder
        consent_version:
          type: string
          example: "2026.1"

    OnboardingResponse:
      type: object
      required:
        - patient
        - consent
      properties:
        patient:
          $ref: '#/components/schemas/Patient'
        consent:
          $ref: '#/components/schemas/Consent'

    GameEventInput:
      type: object
      required:
        - id
        - patient_id
        - session_id
        - domain
        - item_id
        - difficulty
        - correct
        - response_time_ms
        - timestamp
      properties:
        id:
          type: string
          description: Client-generated UUIDv7
          example: 018f3a2b-8c4d-7123-8abc-0123456789ab
        patient_id:
          type: string
        session_id:
          type: string
        domain:
          type: string
          example: memory_match
        item_id:
          type: string
          example: card-bihu-01
        difficulty:
          type: number
          format: float
          example: -0.5
        correct:
          type: boolean
        response_time_ms:
          type: integer
          example: 2450
        timestamp:
          type: string
          format: date-time

    BatchEventsRequest:
      type: object
      required:
        - patient_id
        - events
      properties:
        patient_id:
          type: string
        events:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/GameEventInput'

    BatchEventsResponse:
      type: object
      required:
        - acknowledged_ids
        - new_theta
        - processed_count
      properties:
        acknowledged_ids:
          type: array
          items:
            type: string
        new_theta:
          type: number
          format: float
          example: 0.32
        processed_count:
          type: integer
          example: 10

    AbilityScore:
      type: object
      required:
        - id
        - patient_id
        - domain
        - theta
        - recorded_at
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        domain:
          type: string
          example: memory_match
        theta:
          type: number
          format: float
          example: 0.32
        recorded_at:
          type: string
          format: date-time

    MemoryCreate:
      type: object
      required:
        - memory_type
        - caption
      properties:
        memory_type:
          type: string
          enum: [photo, song, story]
          example: photo
        caption:
          type: string
          example: Rongali Bihu festival at ancestral home
        people:
          type: array
          items:
            type: string
          example: ["Grandmother", "Ranjit"]
        year:
          type: integer
          example: 1985
        file_path:
          type: string
          example: photos/bihu_1985.jpg

    Memory:
      type: object
      required:
        - id
        - patient_id
        - memory_type
        - caption
        - storage_path
        - signed_url
        - created_at
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        memory_type:
          type: string
          enum: [photo, song, story]
        caption:
          type: string
        people:
          type: array
          items:
            type: string
        year:
          type: integer
        storage_bucket:
          type: string
          example: memora-private-memories
        storage_path:
          type: string
        signed_url:
          type: string
          description: Time-limited signed URL for private bucket access
        embedding_id:
          type: string
          format: uuid
        is_approved:
          type: boolean
        created_at:
          type: string
          format: date-time

    QuizOption:
      type: object
      required:
        - id
        - text
        - is_correct
      properties:
        id:
          type: string
        text:
          type: string
        is_correct:
          type: boolean

    GeneratedQuizItem:
      type: object
      required:
        - id
        - patient_id
        - memory_id
        - question_text
        - options
        - difficulty
        - domain
        - approval_status
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        memory_id:
          type: string
          format: uuid
        question_text:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuizOption'
        difficulty:
          type: number
          format: float
          example: 0.0
        domain:
          type: string
          example: reminiscence
        image_url:
          type: string
        approval_status:
          type: string
          enum: [pending, approved, rejected]
          example: pending
        is_approved:
          type: boolean
          example: false

    AshaPatientSummary:
      type: object
      required:
        - id
        - full_name
        - status
        - status_label
        - triage_priority
        - unwell_today
        - is_synthetic
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        status:
          type: string
          enum: [checkin_suggested, watch, steady]
          example: checkin_suggested
        status_label:
          type: string
          example: Check-in suggested
        triage_priority:
          type: integer
          example: 1
        last_active:
          type: string
        sessions_this_week:
          type: integer
          example: 5
        average_score:
          type: number
          format: float
          example: 0.78
        active_alert_domain:
          type: string
        unwell_today:
          type: boolean
          example: false
        is_synthetic:
          type: boolean
          example: true

    AshaPatientDetail:
      type: object
      required:
        - id
        - full_name
        - preferred_language
        - status
        - status_label
        - triage_priority
        - unwell_today
        - is_synthetic
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        preferred_language:
          type: string
          example: as
        status:
          type: string
          enum: [checkin_suggested, watch, steady]
        status_label:
          type: string
        triage_priority:
          type: integer
        unwell_today:
          type: boolean
        is_synthetic:
          type: boolean
        guardian_name:
          type: string
        guardian_relationship:
          type: string
        current_theta:
          type: number
          format: float
        recent_domains:
          type: array
          items:
            type: object
        active_alerts:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time